Extract writer activation service
